update deps and styles

+++ .stylelintrc.js
@@ -2,7 +2,7 @@ module.exports = {
   'extends': 'stylelint-config-standard',
   'rules': {
     'no-empty-source': null,
-    'string-quotes': 'double',
+    'string-quotes': 'single',
     'at-rule-no-unknown': [
       true,
       {
